// src/pages/dashboard/types.ts
export type DashboardType = {
    percentualTotalClientesAlta: number
    percentualTotalClientesBaixa: number
    percentualTotalProdutosAlta: number
    percentualTotalProdutosBaixa: number
    percentualVariacaoClientesAlta: number
    percentualVariacaoClientesBaixa: number
    percentualVariacaoProdutosAlta: number
    percentualVariacaoProdutosBaixa: number
    quantidadeClientesAlta: number
    quantidadeClientesBaixa: number
    quantidadeProdutosAlta: number
    quantidadeProdutosBaixa: number
  }